Include vs files, new filesystem, new typeid
Included vs files in the git to make branching easier. Modified the way
the filesystem to try and fix AssimpIO. Found another fix which would
have made old way work, this one might still be better. Changed the way
the typeID hashing system in the ResourceManager works to try and allow
child classes to exist in their parent's trees.

Removed GLAD submodule to try and add it back using it's "c" branch.